Refactor operation service.go into smaller files (#3808)
* Different files

* Preston's feedback
