+++ beacon-chain/operations/BUILD.bazel
@@ -2,7 +2,14 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "go_default_library",
-    srcs = ["service.go"],
+    srcs = [
+        "attestation.go",
+        "block.go",
+        "exit.go",
+        "log.go",
+        "recent_att_multi_map.go",
+        "service.go",
+    ],
     importpath = "github.com/prysmaticlabs/prysm/beacon-chain/operations",
     visibility = ["//beacon-chain:__subpackages__"],
     deps = [
@@ -30,7 +37,13 @@ go_library(
 go_test(
     name = "go_default_test",
     size = "small",
-    srcs = ["service_test.go"],
+    srcs = [
+        "attestation_test.go",
+        "block_test.go",
+        "exit_test.go",
+        "recent_att_multi_map_test.go",
+        "service_test.go",
+    ],
     embed = [":go_default_library"],
     deps = [
         "//beacon-chain/core/blocks:go_default_library",
