+++ beacon-chain/operations/log.go
@@ -1,0 +1,5 @@
+package operations
+
+import "github.com/sirupsen/logrus"
+
+var log = logrus.WithField("prefix", "operations")
